server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pypdf==4.2.0
python-multipart==0.0.9
aiofiles==24.1.0
